Refactor and only provide results if text is provided
