+++ pages/resources/software/index.vue
@@ -116,7 +116,7 @@ import ResourcesSearchResults from '@/components/Resources/ResourcesSearchResult
 import ToolsAndResourcesFacetMenu from '@/components/FacetMenu/ToolsAndResourcesFacetMenu.vue'
 import { fetchResources, searchTypes, sortOptions } from '../utils.ts'
 import SubmitToolSection from '@/components/Resources/SubmitToolSection.vue'
-import AlternativeSearchResults from '@/components/AlternativeSearchResults/AlternativeSearchResults.vue'
+import AlternativeSearchResults from '~/components/AlternativeSearchResults/AlternativeSearchResultsResources.vue'
 
 export default {
   name: 'SoftwarePage',
